+++ R/000-wrappers.R
@@ -1223,6 +1223,13 @@ class(`PlRDataTypeExpr`) <- c("PlRDataTypeExpr__bundle", "savvy_polars__sealed")
   }
 }
 
+`PlRExpr_arr_eval` <- function(self) {
+  function(`expr`, `as_list`) {
+    `expr` <- .savvy_extract_ptr(`expr`, "PlRExpr")
+    .savvy_wrap_PlRExpr(.Call(savvy_PlRExpr_arr_eval__impl, `self`, `expr`, `as_list`))
+  }
+}
+
 `PlRExpr_arr_get` <- function(self) {
   function(`index`, `null_on_oob`) {
     `index` <- .savvy_extract_ptr(`index`, "PlRExpr")
@@ -3442,6 +3449,7 @@ class(`PlRDataTypeExpr`) <- c("PlRDataTypeExpr__bundle", "savvy_polars__sealed")
   e$`arr_arg_min` <- `PlRExpr_arr_arg_min`(ptr)
   e$`arr_contains` <- `PlRExpr_arr_contains`(ptr)
   e$`arr_count_matches` <- `PlRExpr_arr_count_matches`(ptr)
+  e$`arr_eval` <- `PlRExpr_arr_eval`(ptr)
   e$`arr_get` <- `PlRExpr_arr_get`(ptr)
   e$`arr_join` <- `PlRExpr_arr_join`(ptr)
   e$`arr_len` <- `PlRExpr_arr_len`(ptr)